update original apk-utils code

+++ libcui/libcui/addons/libapk-addon/database.c
@@ -263,7 +263,8 @@ void apk_db_dir_unref(struct apk_database *db, struct apk_db_dir *dir, int allow
 		dir->uid = (uid_t) -1;
 		dir->gid = (gid_t) -1;
 
-		unlinkat(db->root_fd, dir->name, AT_REMOVEDIR);
+		if (dir->namelen)
+			unlinkat(db->root_fd, dir->name, AT_REMOVEDIR);
 	} else if (dir->recalc_mode) {
 		/* Directory permissions need a reset. */
 		apk_db_dir_mkdir(db, dir);
@@ -529,6 +530,11 @@ struct apk_package *apk_db_pkg_add(struct apk_database *db, struct apk_package *
 
 	if (pkg->license == NULL)
 		pkg->license = apk_blob_atomize(APK_BLOB_NULL);
+
+	/* Set as "cached" if installing from specified file, and
+	 * for virtual packages */
+	if (pkg->filename != NULL || pkg->installed_size == 0)
+		pkg->repos |= BIT(APK_REPOSITORY_CACHED);
 
 	idb = apk_hash_get(&db->available.packages, APK_BLOB_CSUM(pkg->csum));
 	if (idb == NULL) {
@@ -868,11 +874,11 @@ static int apk_db_write_fdb(struct apk_database *db, struct apk_ostream *os)
 				apk_blob_push_blob(&bbuf, APK_BLOB_STR("\n"));
 
 				if (os->write(os, buf, bbuf.ptr - buf) != bbuf.ptr - buf)
-					return -1;
+					return -EIO;
 				bbuf = APK_BLOB_BUF(buf);
 			}
 			if (os->write(os, buf, bbuf.ptr - buf) != bbuf.ptr - buf)
-				return -1;
+				return -EIO;
 			bbuf = APK_BLOB_BUF(buf);
 		}
 		os->write(os, "\n", 1);
@@ -1095,7 +1101,7 @@ static int write_index_entry(apk_hash_item item, void *ctx)
 		return r;
 
 	if (iwctx->os->write(iwctx->os, "\n", 1) != 1)
-		return -1;
+		return -EIO;
 
 	iwctx->count++;
 	return 0;
@@ -1118,12 +1124,12 @@ static int apk_db_index_write_nr_cache(struct apk_database *db)
 				 "installed.new",
 				 0644);
 	if (os == NULL)
-		return -1;
+		return -EIO;
 
 	ctx.os = os;
 	list_for_each_entry(ipkg, &db->installed.packages, installed_pkgs_list) {
 		struct apk_package *pkg = ipkg->pkg;
-		if (pkg->repos != 0 || !pkg->in_cache)
+		if (pkg->repos != BIT(APK_REPOSITORY_CACHED))
 			continue;
 		r = write_index_entry(pkg, &ctx);
 		if (r != 0)
@@ -1227,7 +1233,7 @@ static int apk_db_create(struct apk_database *db)
 
 	mkdirat(db->root_fd, "tmp", 01777);
 	mkdirat(db->root_fd, "dev", 0755);
-	mknodat(db->root_fd, "dev/null", 0666, makedev(1, 3));
+	mknodat(db->root_fd, "dev/null", S_IFCHR | 0666, makedev(1, 3));
 	mkdirat(db->root_fd, "etc", 0755);
 	mkdirat(db->root_fd, "etc/apk", 0755);
 	mkdirat(db->root_fd, "lib", 0755);
@@ -1325,7 +1331,7 @@ static void mark_in_cache(struct apk_database *db, int dirfd, const char *name, 
 	if (pkg == NULL)
 		return;
 
-	pkg->in_cache = 1;
+	pkg->repos |= BIT(APK_REPOSITORY_CACHED);
 }
 
 static int add_repos_from_file(void *ctx, int dirfd, const char *file)
@@ -1347,6 +1353,27 @@ static int add_repos_from_file(void *ctx, int dirfd, const char *file)
 	free(blob.ptr);
 
 	return 0;
+}
+
+static void apk_db_setup_repositories(struct apk_database *db)
+{
+	int repo_tag;
+
+	db->repos[APK_REPOSITORY_CACHED] = (struct apk_repository) {
+		.url = "cache",
+		.csum.data = {
+			0xb0,0x35,0x92,0x80,0x6e,0xfa,0xbf,0xee,0xb7,0x09,
+			0xf5,0xa7,0x0a,0x7c,0x17,0x26,0x69,0xb0,0x05,0x38 },
+		.csum.type = APK_CHECKSUM_SHA1,
+	};
+
+	db->num_repos = APK_REPOSITORY_FIRST_CONFIGURED;
+	db->local_repos |= BIT(APK_REPOSITORY_CACHED);
+	db->available_repos |= BIT(APK_REPOSITORY_CACHED);
+
+	/* Get first repository tag (the NULL tag) */
+	repo_tag = apk_db_get_tag_id(db, APK_BLOB_NULL);
+	db->repo_tags[repo_tag].allowed_repos |= BIT(APK_REPOSITORY_CACHED);
 }
 
 int apk_db_open(struct apk_database *db, struct apk_db_options *dbopts)
@@ -1380,8 +1407,7 @@ int apk_db_open(struct apk_database *db, struct apk_db_options *dbopts)
 	apk_protected_path_array_init(&db->protected_paths);
 	db->permanent = 1;
 
-	/* Get first repository tag (the NULL tag) */
-	apk_db_get_tag_id(db, APK_BLOB_NULL);
+	apk_db_setup_repositories(db);
 
 	db->root = strdup(dbopts->root ?: "/");
 	db->root_fd = openat(AT_FDCWD, db->root, O_RDONLY | O_CLOEXEC);
@@ -1651,7 +1677,7 @@ void apk_db_close(struct apk_database *db)
 		}
 	}
 
-	for (i = 0; i < db->num_repos; i++) {
+	for (i = APK_REPOSITORY_FIRST_CONFIGURED; i < db->num_repos; i++) {
 		free(db->repos[i].url);
 		free(db->repos[i].description.ptr);
 	}
@@ -1847,8 +1873,10 @@ struct apk_package *apk_db_get_file_owner(struct apk_database *db,
 	if (filename.len && filename.ptr[0] == '/')
 		filename.len--, filename.ptr++;
 
-	if (!apk_blob_rsplit(filename, '/', &key.dirname, &key.filename))
-		return NULL;
+	if (!apk_blob_rsplit(filename, '/', &key.dirname, &key.filename)) {
+		key.dirname = APK_BLOB_NULL;
+		key.filename = filename;
+	}
 
 	dbf = (struct apk_db_file *) apk_hash_get(&db->installed.files,
 						  APK_BLOB_BUF(&key));
@@ -1876,11 +1904,6 @@ int apk_repo_format_filename(char *buf, size_t len,
 	return n;
 }
 
-int apk_repo_is_remote(struct apk_repository *repo)
-{
-	return repo->csum.type != APK_CHECKSUM_NONE;
-}
-
 static struct apk_bstream *apk_repo_file_open(struct apk_repository *repo,
 					      apk_blob_t *arch,
 					      const char *file,
@@ -1889,9 +1912,6 @@ static struct apk_bstream *apk_repo_file_open(struct apk_repository *repo,
 	const char *url = repo->url;
 
 	apk_repo_format_filename(buf, buflen, url, arch, file);
-
-	if ((apk_flags & APK_NO_NETWORK) && apk_repo_is_remote(repo))
-		return NULL;
 
 	/* We should not get called for non-repository files */
 	if (strcmp(repo->url, "cache") == 0)
@@ -1903,48 +1923,29 @@ static struct apk_bstream *apk_repo_file_open(struct apk_repository *repo,
 struct apk_repository *apk_db_select_repo(struct apk_database *db,
 					  struct apk_package *pkg)
 {
-	static struct apk_repository cache_repo = {
-		.url = "cache",
-		.csum.data = {
-			0xb0,0x35,0x92,0x80,0x6e,0xfa,0xbf,0xee,0xb7,0x09,
-			0xf5,0xa7,0x0a,0x7c,0x17,0x26,0x69,0xb0,0x05,0x38 },
-		.csum.type = APK_CHECKSUM_SHA1,
-	};
-	unsigned int repos = pkg->repos & ~(db->bad_repos);
+	unsigned int repos;
 	int i;
 
-	/* Uninstalled and unavailable? */
-	if (repos == 0 && pkg->ipkg == NULL)
+	/* Select repositories to use */
+	repos = pkg->repos & db->available_repos;
+	if (repos == 0)
 		return NULL;
 
-	if ((repos & db->local_repos) == 0) {
-		/* Network repository (or installed and unavailable) */
-		if (apk_flags & APK_NO_NETWORK) {
-			if (pkg->in_cache)
-				return &cache_repo;
-			return NULL;
-		}
-	} else {
-		/* Local repositories; don't use network */
+	if (repos & db->local_repos)
 		repos &= db->local_repos;
-	}
 
 	/* Pick first repository providing this package */
-	for (i = 0; i < APK_MAX_REPOS; i++) {
+	for (i = APK_REPOSITORY_FIRST_CONFIGURED; i < APK_MAX_REPOS; i++) {
 		if (repos & BIT(i))
 			return &db->repos[i];
 	}
-
-	return NULL;
+	return &db->repos[APK_REPOSITORY_CACHED];
 }
 
-int apk_repository_update(struct apk_database *db, struct apk_repository *repo)
+static int apk_repository_update(struct apk_database *db, struct apk_repository *repo)
 {
 	char cacheitem[PATH_MAX];
 	int r;
-
-	if (!apk_repo_is_remote(repo))
-		return 0;
 
 	apk_cache_format_index(APK_BLOB_BUF(cacheitem), repo);
 	r = apk_cache_download(db, repo->url, db->arch, apkindex_tar_gz, cacheitem,
@@ -2052,7 +2053,7 @@ int apk_db_add_repository(apk_database_t _db, apk_blob_t _repository)
 		repo = &db->repos[repo_num];
 		if (strcmp(url, repo->url) == 0) {
 			db->repo_tags[tag_id].allowed_repos |=
-				BIT(repo_num) & ~db->bad_repos;
+				BIT(repo_num) & db->available_repos;
 			free(url);
 			return 0;
 		}
@@ -2068,29 +2069,32 @@ int apk_db_add_repository(apk_database_t _db, apk_blob_t _repository)
 		.url = url,
 	};
 
+	apk_blob_checksum(brepo, apk_checksum_default(), &repo->csum);
+
 	if (apk_url_local_file(repo->url) == NULL) {
-		apk_blob_checksum(brepo, apk_checksum_default(), &repo->csum);
-
-		if (apk_flags & APK_UPDATE_CACHE)
-			apk_repository_update(db, repo);
-
+		if (!(apk_flags & APK_NO_NETWORK)) {
+			db->available_repos |= BIT(repo_num);
+			if (apk_flags & APK_UPDATE_CACHE)
+				apk_repository_update(db, repo);
+		}
 		apk_cache_format_index(APK_BLOB_BUF(buf), repo);
 		bs = apk_bstream_from_file(db->cache_fd, buf);
 	} else {
 		db->local_repos |= BIT(repo_num);
+		db->available_repos |= BIT(repo_num);
 		bs = apk_repo_file_open(repo, db->arch, apkindex_tar_gz, buf, sizeof(buf));
 	}
-	if (bs != NULL) {
+	if (bs != NULL)
 		r = load_index(db, bs, targz, repo_num);
-	} else
+	else
 		r = -ENOENT;
 
 	if (r != 0) {
 		apk_warning("Ignoring %s: %s", buf, apk_error_str(r));
-		db->bad_repos |= BIT(repo_num);
+		db->available_repos &= ~BIT(repo_num);
 		r = 0;
 	} else {
-		db->repo_tags[tag_id].allowed_repos |= BIT(repo_num);
+		db->repo_tags[tag_id].allowed_repos |= BIT(repo_num) & db->available_repos;
 	}
 
 	return 0;
@@ -2161,6 +2165,21 @@ static int read_info_line(void *_ctx, apk_blob_t line)
 	return 0;
 }
 
+static struct apk_db_dir_instance *apk_db_install_directory_entry(struct install_ctx * ctx, apk_blob_t dir)
+{
+	struct apk_database *db = ctx->db;
+	struct apk_package *pkg = ctx->pkg;
+	struct apk_installed_package *ipkg = pkg->ipkg;
+	struct apk_db_dir_instance *diri;
+
+	if (ctx->diri_node == NULL)
+		ctx->diri_node = hlist_tail_ptr(&ipkg->owned_dirs);
+	ctx->diri = diri = apk_db_diri_new(db, pkg, dir, &ctx->diri_node);
+	ctx->file_diri_node = hlist_tail_ptr(&diri->owned_files);
+
+	return diri;
+}
+
 static int apk_db_install_archive_entry(void *_ctx,
 					const struct apk_file_info *ae,
 					struct apk_istream *is)
@@ -2215,8 +2234,10 @@ static int apk_db_install_archive_entry(void *_ctx,
 	/* Installable entry */
 	ctx->current_file_size = apk_calc_installed_size(ae->size);
 	if (!S_ISDIR(ae->mode)) {
-		if (!apk_blob_rsplit(name, '/', &bdir, &bfile))
-			return 0;
+		if (!apk_blob_rsplit(name, '/', &bdir, &bfile)) {
+			bdir = APK_BLOB_NULL;
+			bfile = name;
+		}
 
 		if (bfile.len > 6 && memcmp(bfile.ptr, ".keep_", 6) == 0)
 			return 0;
@@ -2224,9 +2245,12 @@ static int apk_db_install_archive_entry(void *_ctx,
 		/* Make sure the file is part of the cached directory tree */
 		diri = find_diri(ipkg, bdir, diri, &ctx->file_diri_node);
 		if (diri == NULL) {
-			apk_error("%s: File '%*s' entry without directory entry.\n",
-				  pkg->name->name, name.len, name.ptr);
-			return -1;
+			if (!APK_BLOB_IS_NULL(bdir)) {
+				apk_error("%s: File '%*s' entry without directory entry.\n",
+					  pkg->name->name, name.len, name.ptr);
+				return -1;
+			}
+			diri = apk_db_install_directory_entry(ctx, bdir);
 		}
 
 		opkg = NULL;
@@ -2329,12 +2353,7 @@ static int apk_db_install_archive_entry(void *_ctx,
 		if (name.ptr[name.len-1] == '/')
 			name.len--;
 
-		if (ctx->diri_node == NULL)
-			ctx->diri_node = hlist_tail_ptr(&ipkg->owned_dirs);
-		ctx->diri = diri = apk_db_diri_new(db, pkg, name,
-						   &ctx->diri_node);
-		ctx->file_diri_node = hlist_tail_ptr(&diri->owned_files);
-
+		diri = apk_db_install_directory_entry(ctx, name);
 		apk_db_diri_set(diri, ae->mode, ae->uid, ae->gid);
 		apk_db_dir_mkdir(db, diri->dir);
 	}
@@ -2360,8 +2379,10 @@ static void apk_db_purge_pkg(struct apk_database *db,
 			diri->dir->modified = 1;
 
 		hlist_for_each_entry_safe(file, fc, fn, &diri->owned_files, diri_files_list) {
-			snprintf(name, sizeof(name), "%s/%s%s",
-				 diri->dir->name, file->name, exten ?: "");
+			snprintf(name, sizeof(name),
+				 PKG_FILE_FMT "%s",
+				 PKG_FILE_PRINTF(diri->dir, file),
+				 exten ?: "");
 
 			key = (struct apk_db_file_hash_key) {
 				.dirname = APK_BLOB_PTR_LEN(diri->dir->name, diri->dir->namelen),
@@ -2406,10 +2427,10 @@ static void apk_db_migrate_files(struct apk_database *db,
 		dir->modified = 1;
 
 		hlist_for_each_entry_safe(file, fc, fn, &diri->owned_files, diri_files_list) {
-			snprintf(name, sizeof(name), "%s/%s",
-				 diri->dir->name, file->name);
-			snprintf(tmpname, sizeof(tmpname), "%s/%s.apk-new",
-				 diri->dir->name, file->name);
+			snprintf(name, sizeof(name), PKG_FILE_FMT,
+				 PKG_FILE_PRINTF(diri->dir, file));
+			snprintf(tmpname, sizeof(tmpname), PKG_FILE_FMT ".apk-new",
+				 PKG_FILE_PRINTF(diri->dir, file));
 
 			key = (struct apk_db_file_hash_key) {
 				.dirname = APK_BLOB_PTR_LEN(dir->name, dir->namelen),
@@ -2497,16 +2518,14 @@ static int apk_db_unpack_pkg(struct apk_database *db,
 			return -1;
 		}
 
-		if (apk_db_cache_active(db) && apk_repo_is_remote(repo)) {
+		if (strcmp(repo->url, "cache") == 0) {
 			apk_pkg_format_cache(pkg, APK_BLOB_BUF(file));
 			bs = apk_bstream_from_file(db->cache_fd, file);
-		}
-
-		if (bs == NULL) {
+		} else {
 			apk_pkg_format_plain(pkg, APK_BLOB_BUF(item));
 			bs = apk_repo_file_open(repo, pkg->arch,
 						item, file, sizeof(file));
-			if (apk_repo_is_remote(repo))
+			if (!(pkg->repos & db->local_repos))
 				need_copy = TRUE;
 		}
 	} else {
@@ -2550,7 +2569,7 @@ static int apk_db_unpack_pkg(struct apk_database *db,
 	if (need_copy) {
 		if (r == 0) {
 			renameat(db->cachetmp_fd, item, db->cache_fd, item);
-			pkg->in_cache = 1;
+			pkg->repos |= BIT(APK_REPOSITORY_CACHED);
 		} else {
 			unlinkat(db->cachetmp_fd, item, 0);
 		}
